+++ examples/common/mod.rs
@@ -94,8 +94,8 @@ pub fn drop_table(
     let sql = format!("drop table {table_name} purge");
     let result = conn.execute(&sql, &[]);
     if let Err(err) = result
-        && let oracledb::ErrorKind::DbError(message) = err.kind()
-        && !message.starts_with("ORA-00942:")
+        && let oracledb::ErrorKind::DbError(db_error) = err.kind()
+        && db_error.code() != 942
     {
         return Err(err);
     }