+++ doc/release_notes.md
@@ -10,6 +10,9 @@
     [Metadata::vector_dimensions()](crate::Metadata::vector_dimensions()) and
     [Metadata::vector_dimensions()](crate::Metadata::vector_storage_format())
     and the enumeration [VectorStorageFormat](crate::VectorStorageFormat).
+1.  Added struct [DbError](crate::DbError) containing information about the
+    database error which is now returned instead of `String` for
+    the [ErrorKind::DbError](crate::ErrorKind::DbError) enum variant.
 1.  Eliminated hang when an error occurs during a DML returning statemnt
     ([issue 16](https://github.com/oracle/rust-oracledb/issues/16)).
 1.  Ensure that connections returned from a pool always start with a call